move `react-connect-kit` types to `types` pkg
